cleaned up styles, added styles to grid gallery
