+++ src/components/Tags/TagSingle.tsx
@@ -1,8 +1,8 @@
 import * as React from "react";
-import { Container, Wrapper, Header } from "./TagListStyles";
+import * as Tags from "../../styles/AppStyles";
 import * as actions from "../../actions";
 import { connect } from "react-redux";
-import baseUrl from '../../api/baseurl'
+import baseUrl from "../../api/baseurl";
 import Gallery from "react-grid-gallery";
 
 class TagSingle extends React.Component<any, any> {
@@ -10,23 +10,33 @@ class TagSingle extends React.Component<any, any> {
     this.props.fetchImages(this.props.match.params.tagname);
   }
   renderImages() {
-    console.log(this.props.images)
+    console.log(this.props.images);
     return this.props.images.map(tag => {
-      return <img src={baseUrl + '/photos/image/' + tag.imageId} alt="image" width='40%' height='40%' />;
+      return (
+        <img
+          src={baseUrl + "/photos/image/" + tag.imageId}
+          alt="image"
+          width="40%"
+          height="40%"
+        />
+      );
     });
   }
 
   public render() {
-
-    console.log(this.props)
+    console.log(this.props);
     const images = [];
-    this.props.images.forEach(function (tag) {
+    this.props.images.forEach(function(tag) {
       images.push({
         src: baseUrl + "/photos/image/" + tag.imageId,
         thumbnail: baseUrl + "/photos/image/" + tag.imageId,
-        thumbnailWidth: '60%',
-        thumbnailHeight: '60%',
-        caption: ['Model: ' + tag.source.model + ' | ', 'Confidence: ' + (tag.confidence * 100).toFixed(2) + '% | ', 'Date: ' + tag.dateAdded],
+        thumbnailWidth: "60%",
+        thumbnailHeight: "60%",
+        caption: [
+          "Model: " + tag.source.model + " | ",
+          "Confidence: " + (tag.confidence * 100).toFixed(2) + "% | ",
+          "Date: " + tag.dateAdded
+        ],
         tags: [{ value: tag.source.type }]
       });
     });
@@ -41,16 +51,16 @@ class TagSingle extends React.Component<any, any> {
     };
 
     return (
-      <Container>
-        <Wrapper>
-          <Header>
+      <Tags.Container>
+        <Tags.Wrapper>
+          <Tags.Header>
             <h2>{this.props.match.params.tagname}</h2>
-          </Header>
+          </Tags.Header>
           <div>
             <Gallery images={images} backdropClosesModal tagStyle={tagStyles} />
           </div>
-        </Wrapper>
-      </Container>
+        </Tags.Wrapper>
+      </Tags.Container>
     );
   }
 }
