+++ src/index.tsx
@@ -3,19 +3,20 @@ import ReactDOM from "react-dom";
 import { createStore, applyMiddleware } from "redux";
 import { Provider } from "react-redux";
 import reducers from "./reducers";
-import App from "./components/App";
+import AppRouter from "./routes/AppRouter";
 import thunk from "redux-thunk";
 
-// STYLES
+// Styles
+import "normalize.css/normalize.css";
 import "@blueprintjs/core/lib/css/blueprint.css";
-import "./scss/base.scss";
+import "./styles/CustomStyles.scss";
 
-// REDUX STORE
+// Redux Config
 const store = createStore(reducers, applyMiddleware(thunk));
 
 ReactDOM.render(
   <Provider store={store}>
-    <App />
+    <AppRouter />
   </Provider>,
   document.getElementById("root")
 );
