--- src/scss/base.scss
@@ -1,72 +1,0 @@
-// CSS RESET
-* {
-  margin: 0;
-  padding: 0;
-}
-
-*,
-*::after,
-*::before {
-  -webkit-box-sizing: inherit;
-  box-sizing: inherit;
-}
-
-html {
-  -webkit-box-sizing: border-box;
-  box-sizing: border-box;
-  list-style: none !important;
-}
-
-body {
-  text-decoration: none;
-  font-style: none;
-  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;
-}
-
-.navbar__nav-link {
-  width: 100%;
-  min-height: 4rem;
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -webkit-box-pack: center;
-  -ms-flex-pack: center;
-  justify-content: center;
-  -webkit-box-flex: 0;
-  -ms-flex: 0 0 2rem;
-  flex: 0 0 2rem;
-  -webkit-box-align: center;
-  -ms-flex-align: center;
-  align-items: center;
-  -webkit-transition: all 0.2s ease;
-  transition: all 0.2s ease;
-  color: #fff;
-
-  &.selected {
-    background: #48c0b9;
-  }
-
-  & svg {
-    color: #FFF;
-  }
-
-  &:hover,
-  &:active {
-    background: #48c0b9;
-    outline: none;
-
-    & svg {
-      color: #FFF;
-    }
-  }
-}
-
-.pagination {
-  display: flex;
-  justify-content: space-between;
-}
-
-.image-collection {
-  display: flex;
-  justify-content: space-evenly;
-}
