+++ src/styles/CustomStyles.scss
@@ -1,0 +1,22 @@
+body {
+  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;
+}
+
+#ReactGridGallery {
+  display: flex;
+  flex-wrap: wrap;
+
+  .tile {
+    flex: 1;
+    max-width: 33.33%;
+  }
+
+  .tile-viewport {
+    display: flex;
+    flex: 1;
+  }
+
+  img {
+    flex: 1;
+  }
+}
